chore: rename restaurant addresses
